# docs/swagger.yaml
definitions:
  requests.CreateConditionTemplate:
    properties:
      condition:
        maxLength: 2000
        type: string
      title:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - condition
    - title
    type: object
...
  requests.DeleteConditionTemplate:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  requests.EditConditionTemplate:
    properties:
      condition:
        maxLength: 2000
        type: string
      id:
        type: integer
      title:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - condition
    - id
    - title
    type: object
...
  responses.ConditionTemplateDTO:
    properties:
      condition:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  responses.CreateConditionTemplateDTO:
    properties:
      status:
        type: string
      task_template:
        $ref: '#/definitions/responses.ConditionTemplateDTO'
    type: object
...
  responses.DeleteConditionTemplateDTO:
    properties:
      status:
        type: string
    type: object
...
  responses.EditConditionTemplateDTO:
    properties:
      status:
        type: string
      task_template:
        $ref: '#/definitions/responses.ConditionTemplateDTO'
    type: object
...
  responses.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
...
  responses.GetAllConditionTemplatesDTO:
    properties:
      task_templates:
        items:
          $ref: '#/definitions/responses.ConditionTemplateDTO'
        type: array
    type: object
...
  responses.GetConditionTemplateDTO:
    properties:
      task_template:
        $ref: '#/definitions/responses.ConditionTemplateDTO'
    type: object
...
  responses.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/template/condition/all:
    get:
      description: Retrieves all condition templates created by the current user with
        pagination
      parameters:
      - description: 'Pagination offset (default: 0)'
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.GetAllConditionTemplatesDTO'
        "400":
          description: Invalid token or request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get all condition templates for the user
      tags:
      - ConditionTemplate
  /api/template/condition/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a specific condition template owned by the user
      parameters:
      - description: Condition template ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.DeleteConditionTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DeleteConditionTemplateDTO'
        "400":
          description: Invalid token or request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Delete a condition template by ID
      tags:
      - ConditionTemplate
  /api/template/condition/edit:
    put:
      consumes:
      - application/json
      description: Updates a condition template owned by the user
      parameters:
      - description: Condition template data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.EditConditionTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.EditConditionTemplateDTO'
        "400":
          description: Invalid token or request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation errors
          schema:
            $ref: '#/definitions/responses.ValidationErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Edit an existing condition template
      tags:
      - ConditionTemplate
  /api/template/condition/get/{id}:
    get:
      description: Retrieves a specific condition template owned by the user
      parameters:
      - description: Condition Template ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.GetConditionTemplateDTO'
        "400":
          description: Invalid token or request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get a condition template by ID
      tags:
      - ConditionTemplate
  /api/template/condition/new:
    post:
      consumes:
      - application/json
      description: Creates a condition template for the current user
      parameters:
      - description: Condition template data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.CreateConditionTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CreateConditionTemplateDTO'
        "400":
          description: Invalid token or request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation errors
          schema:
            $ref: '#/definitions/responses.ValidationErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Create a new condition template
      tags:
      - ConditionTemplate
...
swagger: "2.0"